Refactoring package and module structure
Define cid_minter as a package
Removed main function from module scripts
Created run scripts for each module scripts
Updated unit tests based on package/module structure
Removed unit tests for the main functions
